fixing really large lightcurve size, addresses #20
end of the lightcurve extraction is the min of the last datapoint
or the exptime.  This fixes issues with bad (high) times in data.

Now there is an additional issue of how to handle the introduced
gaps in the lightcurves.  The events are in the right order, in
the table (and are real) but just have incorrect time labels.